% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{AdjustPValue}
\alias{AdjustPValue}
\title{ALREADY IMPLEMENTED IN rsurvival::AnalyseExpt() and rsurvival::AnalyseSplittedExpt()}
\usage{
AdjustPValue(probs)
}
\description{
ALREADY IMPLEMENTED IN rsurvival::AnalyseExpt() and rsurvival::AnalyseSplittedExpt()
}
