% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{AnalyseSplittedExpt}
\alias{AnalyseSplittedExpt}
\title{A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.}
\usage{
AnalyseSplittedExpt(gt.alive, gt.dead, min.freq.al = 0,
  location.cols = TRUE, deltas = TRUE, genotypic = FALSE,
  prob = TRUE, gt.pop1 = NULL, gt.pop2 = NULL, backup.path = NULL)
}
\arguments{
\item{gt.alive}{A genotype data frame of the alive population}

\item{gt.dead}{A genotype data frame of the dead population}

\item{min.freq.al}{Optional. Numeric from 0 to 1. Variants with MAF under
this threshold will be ommited.}

\item{location.cols}{If TRUE, adds two columns to the result, which are the
scaffold name and the locus name as factors (deafult is FALSE)}

\item{deltas}{If TRUE (default), include columns of differencies between
relative allelic frequencies after selection and before selection in the
result. Positive value means that the allele is more common in the survivers
population.}

\item{genotypic}{If TRUE, will include counts of genotype in the result.}

\item{prob}{If TRUE (default), some probabilities will be included in
the result, as well as frequencies}

\item{gt.pop1}{Optional. A genotype matrix. If both \code{gt.pop1} and \code{gt.pop2}
are set, will include the W&C FST estimation to the result.}

\item{gt.pop2}{Optional. A genotype matrix. If both \code{gt.pop1} and \code{gt.pop2}
are set, will include the W&C FST estimation to the result.}

\item{backup.path}{Optionnal. A path where backup files can be stored}
}
\value{
A data frame of frequencies and probabilities from your experiment.
}
\description{
A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.
}
\examples{
\dontrun{
AnalyseSplittedExpt(gt.alive, gt.dead)
AnalyseSplittedExpt(gt.alive, gt.dead, min.freq.al = 0.1)
AnalyseSplittedExpt(gt.alive, gt.dead, location.cols = FALSE)
AnalyseSplittedExpt(gt.alive, gt.dead, deltas = FALSE, p.values = FALSE)
AnalyseSplittedExpt(gt.alive, gt.dead, backup.path = "example.csv")
}
}
\seealso{
For more information, see \code{\link{CalcProbsSelection}} and
\code{\link{CalcFreqGt}} which this function bind.
}
