% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{Benchmark}
\alias{Benchmark}
\title{A function to perform multiple analysis using the same parameters.}
\usage{
Benchmark(gt, variant.name, s, h, n.dead, n, verbose = TRUE,
  shape = TRUE, include.all = FALSE, max.cores = NULL)
}
\arguments{
\item{gt}{A genotype matrix.}

\item{variant.name}{Name of odded variant (e.g. scaffolf123.987654).}

\item{s}{Fitness parameter s for \code{variant.name}.}

\item{h}{Fitness parameter h for \code{variant.name}.}

\item{n.dead}{Number of dead samples for each iteration.}

\item{n}{Number of iterations. \code{n} analysis will be performed with the
same parameters.}

\item{verbose}{If TRUE (default), info on progression will be printed.}

\item{shape}{If TRUE (default), variants ranking will be returned. Else,
variants will be ordered and returned.}

\item{include.all}{If FALSE (default), only names of the variants will be
returned. Else, all the calculated frequencies and probs will be returned.}

\item{max.cores}{OPTIONAL. Number of cores from the CPU that R can use for
this job. If NULL, all detected cores will be used.}
}
\value{
Ordered variants names or a vector of variants rank. Will return this 
\code{n} times.
}
\description{
A function to perform multiple analysis using the same parameters.
}
\seealso{
\code{\link{BenchmarkScan}} which binds this function.
}
