% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{BenchmarkScan}
\alias{BenchmarkScan}
\title{A function to perform statistical bencharking for a range of parameters.
Launch multiple \code{link{Benchmark}}.}
\usage{
BenchmarkScan(gt, variant.name, surv.ratio = 0.5, s = 2, h = 0.5,
  n.iter = 100, shape = TRUE, include.all = FALSE,
  max.cores = NULL)
}
\arguments{
\item{gt}{A genotype matrix.}

\item{variant.name}{Name of odded variant (e.g. scaffolf123.987654).}

\item{surv.ratio}{Survival ratio from 0 to 1. Could be a vector.}

\item{s}{Fitness parameter s for \code{variant.name}. Could be a vector.}

\item{h}{Fitness parameter h for \code{variant.name}. Could be a vector.}

\item{n.iter}{Number of iterations. \code{n} analysis will be performed for
each combinations of parameters.}

\item{include.all}{If FALSE (default), only names of the variants will be
returned. Else, all the calculated frequencies and probs will be returned.}

\item{max.cores}{OPTIONAL. Number of cores from the CPU that R can use for
this job. If NULL, all detected cores will be used.}

\item{n.dead}{Number of dead samples for each iteration.}
}
\value{
Ordered variants names, calculated \code{n.iter} times for each combinations of
parameters. Only one parameter can variate during a scan. For multiple
values of multiple parameters, please make a loop on top of
\code{link{BenchmarkScan}}.
}
\description{
A function to perform statistical bencharking for a range of parameters.
Launch multiple \code{link{Benchmark}}.
}
\seealso{
\code{\link{BenchmarkScan}} which this function binds.
}
