% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{CalcFreqGt}
\alias{CalcFreqGt}
\title{A function to count occurences of genotypes in a genotype matrix.}
\usage{
CalcFreqGt(gt, genotypic = TRUE, allelic = FALSE, absolute = TRUE,
  percentage = FALSE, totals = TRUE, backup.path = NULL,
  min.freq.gt = NULL, min.freq.al = NULL)
}
\arguments{
\item{gt}{The genotype matrix to analyse}

\item{genotypic}{If TRUE (default), include frequencies of genotypes in the
result}

\item{allelic}{If TRUE, include frequencies of alles in the result,
default set to FALSE}

\item{absolute}{If TRUE (default), occurences will be shown instead of
relative frequencies}

\item{percentage}{If TRUE and absolute is FALSE, show the relative
frequencies in percentage. Default is FALSE.}

\item{totals}{If TRUE (default), returned vector will include totals.}

\item{backup.path}{Optional. Path to the CSV backup file the result must be
saved to.}

\item{min.freq.gt}{Optional. Minimal value of relative frequency for each
genotype in a variant of \code{gt}. Variants which does not fulfill this
condition will be omitted. Default is NULL (disabled).}

\item{min.freq.al}{Optional. Minimal value of relative frequency for each
allele in a variant of \code{gt}. Variants which does not fulfill this
condition will be omitted. Default is 0.1, set it to NULL to disable this
feature.}
}
\value{
A data frame of frequencies.
Each row is a variant and each column is a calculated frequency. 
Column names depend on provided parameters.
Typical column names would be "count_gt_HomoREF" (absolute frequency of homo
REF genotype) or "freq_al_ALT" (relative frequency of ALT alleles).
Column names ending with "MissVal" concern samples with missing values for
this variant.
Column names ending with "Total" concern all samples for this variant,
including the ones with missing data.
}
\description{
A function to count occurences of genotypes in a genotype matrix.
}
\examples{
CalcFreqGt(genotype)
CalcFreqGt(genotype, genotypic = FALSE, allelic = TRUE)
CalcFreqGt(genotype, absolute = FALSE, percentage = TRUE)
CalcFreqGt(genotype, backup.path = "example.csv")
CalcFreqGt(genotype, totals= FALSE, min.freq.al = 0.15)
}
\seealso{
\code{\link{CalcFreqVariant}} which this function bind
}
