% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{CalcFreqMultiVariant}
\alias{CalcFreqMultiVariant}
\title{A function to count occurences of genotypes in multiple combined variant
vectors.}
\usage{
CalcFreqMultiVariant(variants, group.nas = FALSE)
}
\arguments{
\item{variants}{A vector of variants, subset of a genotype matrix.}

\item{group.nas}{If FALSE (default), each combinations with missing values
will be considered on its own. Else, combinations with missing values will
be grouped as one category.}
}
\value{
A vector of frequencies.
Each Item is a calculated frequency for this combination of variants.
}
\description{
A function to count occurences of genotypes in multiple combined variant
vectors.
}
\examples{
\dontrun{
CalcFreqMultiVariant(variants, group.nas = TRUE)
}
}
\seealso{
\code{\link{CalcFreqGt}} which calculate frequencies for a 
genotype matrix.
}
