% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{CalcFreqVariant}
\alias{CalcFreqVariant}
\title{A function to count occurences of genotypes in a variant vector.}
\usage{
CalcFreqVariant(variant, genotypic = TRUE, allelic = FALSE,
  absolute = TRUE, percentage = FALSE, extrapolate.freq = TRUE,
  totals = TRUE, min.freq.gt = 0, min.freq.al = 0,
  include.ids = FALSE)
}
\arguments{
\item{variant}{The genotype matrix to analyse}

\item{genotypic}{If TRUE (default), include frequencies of genotypes in the
result}

\item{allelic}{If TRUE, include frequencies of alles in the result,
default set to FALSE}

\item{absolute}{If TRUE (default), occurences will be shown instead of
relative frequencies}

\item{percentage}{If TRUE and absolute is FALSE, show the relative
frequencies in percentage}

\item{extrapolate.freq}{If TRUE (default), the relative frequencies of the 
signinficant data will be extrapolated. (i.e. freq.al.REF + freq.al.ALT = 1,
even if there are missing datas)}

\item{totals}{If TRUE (default), returned vector will include totals.}

\item{min.freq.gt}{Optional. Minimal value of relative frequency for each
genotype in a \code{variant}. If \code{variant} does not fulfill this
condition, return NULL. Default is NULL (disabled).}

\item{min.freq.al}{Optional. Minimal value of relative frequency for each
allele in a \code{variant}. If \code{variant} does not fulfill this
condition, return NULL. Default is NULL (disabled).}
}
\value{
A vector of frequencies.
Each Item is a calculated frequency for this variant. 
Item names depend on provided parameters.
Typical item names would be "count_gt_HomoREF" (absolute frequency of homo
REF genotype) or "freq_al_ALT" (relative frequency of ALT alleles).
Item names ending with "MissVal" concern samples with missing values for
this variant.
Item names ending with "Total" concern all samples for this variant,
including the ones with missing data.
}
\description{
A function to count occurences of genotypes in a variant vector.
}
\examples{
\dontrun{
CalcFreqVariant(variant)
CalcFreqVariant(variant, genotypic = FALSE, allelic = TRUE)
CalcFreqVariant(variant, absolute = FALSE, percentage = TRUE)
CalcFreqVariant(variant, totals= FALSE, min.freq.al = 0.15)
}
}
\seealso{
\code{\link{CalcFreqGt}} which bind this function
}
