% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{CalcOddedPick}
\alias{CalcOddedPick}
\title{A function to calculate the number of picked lives per categories of samples
according to MWNCHypergeo model.}
\usage{
CalcOddedPick(variant, survival, n = 10, verbose = TRUE)
}
\arguments{
\item{variant}{The odded variant vector}

\item{survival}{A survival vector}

\item{n}{Number of life vectors to be generated (default is 10)}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A list. list$odded.lives is a matrix containing a number of survivors per 
categories (e.g. list$odded.lives row names could be "HOMOREF", "HOMOALT",
...) with \code{n} column, each time generated with the same input values.
list$odded.samples are the index of the samples belonging to each categories
(e.g. list$odded.samples$HETERO could be c(1, 3, 8, 14))
}
\description{
A function to calculate the number of picked lives per categories of samples
according to MWNCHypergeo model.
}
\examples{
CalcOddedPick(variant, survival)
CalcOddedPick(variant, survival, n = 1000)
CalcOddedPick(variant, survival, verbose = FALSE)
}
\seealso{
For more detail, see \code{\link{IterRandPick}} which binds this
function
}
