% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{CalcProbsSelection}
\alias{CalcProbsSelection}
\title{A function to calculate probabilities that observed frequencies are due to
positive selection using MWNCHHypergeo model.}
\usage{
CalcProbsSelection(freq.alive, freq.all, map.alive = NULL,
  map.all = NULL)
}
\arguments{
\item{freq.alive}{An vector of observed absolute frequencies for the three
genotypes in the alive population}

\item{freq.all}{An vector of observed absolute frequencies for the three
genotypes in the population before selection}
}
\value{
A vector of probabilities : \code{p_neutral} is the probability to obtain
this distribution with a random picking process, \code{p_select} is the
probability to obtain this distribution with a picking in a biased urn,
\code{weights} are the odds used to calculate \code{p_select}, \code{lrt}
is the result of the likelyhood test of \code{p_select} uppon
\code{p_neutral}, \code{p_value} is the probability that the distribution
is due to positive selection.
}
\description{
A function to calculate probabilities that observed frequencies are due to
positive selection using MWNCHHypergeo model.
}
\examples{
f.alive <- c(17, 3, 15, 1, 36)
f.all   <- c(24, 5, 22, 2, 53)
CalcProbsSelection(f.alive, f.all)
}
\seealso{
For more information, see \code{\link{CalcWeightsSurvival}} wich
this function bind.
}
