% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{CalcWeightsPredation}
\alias{CalcWeightsPredation}
\title{A function to calculate odds for predation from observed selection
frequencies using MWNCHHypergeo model.}
\usage{
CalcWeightsPredation(freq.alive, freq.all, map.alive = NULL,
  map.all = NULL)
}
\arguments{
\item{freq.alive}{A vector of observed absolute frequencies for the three
genotypes in the alive population}

\item{freq.all}{An vector of observed absolute frequencies for the three
genotypes in the population before selection}

\item{map.alive}{Optional. Integer list containing the ids of specific
columns of \code{freq.alive}. Increase speed for large scale calls. Items of
\code{map.alive} must be named "homo.ref", "hetero", "homo.alt", "missval",
"total".}

\item{map.all}{Optional. Integer list containing the ids of specific
columns of \code{freq.all}. Increase speed for large scale calls. Items of
\code{map.all} must be named "homo.ref", "hetero", "homo.alt", "missval",
"total".}
}
\value{
A vector of odds for the 3 genotypes, according to the MWNCHHypergeo model.
}
\description{
A function to calculate odds for predation from observed selection
frequencies using MWNCHHypergeo model.
}
\examples{
\dontrun{
f.alive <- c(17, 3, 15, 1, 36)
f.all  <- c(24, 5, 22, 2, 53)
CalcWeightsPredation(f.alive, f.all)
}
}
\seealso{
For more information, see \code{\link{FindIdsGtCounts}} which this
function binds.
}
