% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{CalcWeightsSelection}
\alias{CalcWeightsSelection}
\title{A function to calculate odds for positive selection from observed selection
frequencies using MWNCHHypergeo model.}
\usage{
CalcWeightsSelection(freq.dead, freq.all)
}
\arguments{
\item{freq.dead}{An vector of observed absolute frequencies for the three
genotypes in the dead population}

\item{freq.dead}{An vector of observed absolute frequencies for the three
genotypes in the population before selection}
}
\value{
A vector of odds for the 3 genotypes, according to the MWNCHHypergeo model.
}
\description{
A function to calculate odds for positive selection from observed selection
frequencies using MWNCHHypergeo model.
}
\examples{
f.dead <- c(17, 3, 15, 1, 36)
f.all  <- c(24, 5, 22, 2, 53)
CalcWeightsSelection(f.dead, f.all)
}
