% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{CensorLowDepth}
\alias{CensorLowDepth}
\title{A function to replace genotype data with low depth by NA}
\usage{
CensorLowDepth(gt, dp, min.depth, verbose = TRUE)
}
\arguments{
\item{gt}{A dataframe to process (e.g. a genotype dataframe)}

\item{dp}{A dataframe from which the depth of \code{gt} will be extracted
(e.g. a depth dataframe)}

\item{min.depth}{Integer. The minimum depth of each \code{gt} cells, under wich the
data will be replaced by NA}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution}
}
\value{
A list of dataframes. 
\code{list$gt} is \code{gt} with NA instead of the data with low depth.
\code{list$dp} is \code{dp} with NA instead of the low values of depth.
}
\description{
A function to replace genotype data with low depth by NA
}
\examples{
censorLowDepth(genotype, depth, 4)
censorLowDepth(genotype, depth, 4, verbose = FALSE)
}
