% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ConvertGtToFactors}
\alias{ConvertGtToFactors}
\title{A function to convert genotype character data frame to variant factor data
frame.}
\usage{
ConvertGtToFactors(gt)
}
\arguments{
\item{gt}{a genotype character data frame.}
}
\value{
A genotype data frame in which each genotype is a factor level.
Supported codes are: "0/0", "0|0", "1/0", "1|0", "0/1", "0|1", "1/1", "1|1".
}
\description{
A function to convert genotype character data frame to variant factor data
frame.
}
\examples{
\dontrun{
 ConvertGtToFactors(gt)
}
}
\seealso{
For a single variant data frame see
\code{\link{ConvertVariantToFactors}}
}
