% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ConvertVariantToFactors}
\alias{ConvertVariantToFactors}
\title{A function to convert variant character data frame to variant factor data
frame.}
\usage{
ConvertVariantToFactors(variant)
}
\arguments{
\item{variant}{a variant character data frame.}
}
\value{
A variant data frame in which each genotype is a factor level.
Supported codes are: "0/0", "0|0", "1/0", "1|0", "0/1", "0|1", "1/1", "1|1".
}
\description{
A function to convert variant character data frame to variant factor data
frame.
}
\examples{
\dontrun{
 ConvertVariantToFactors (variant)
}
}
\seealso{
For a whole genotype data frame see
\code{\link{ConvertGtToFactors}}
}
