% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{CountSignSnps}
\alias{CountSignSnps}
\title{A function to count snps with a low probability of being due to a neutral
model}
\usage{
CountSignSnps(probs, max.p.neutral = 0.01)
}
\arguments{
\item{probs}{A probability data frame}

\item{max.p.neutral}{A threshold for \code{p.neutral} under which the snp
will be counted (default is 0.01)}
}
\value{
An integer. The number of snps with low \code{p.neutral} value in
\code{probs} data frame.
}
\description{
A function to count snps with a low probability of being due to a neutral
model
}
\examples{
\dontrun{
CountSignSnps(probs)
CountSignSnps(probs, max.p.neutral = 0.05)
}
}
\seealso{
For more detail, see \code{\link{AnalyseSplittedExpt}} or
\code{\link{CalcProbsSelection}} which can provide you a probs data frame.
}
