% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{ExtractGt}
\alias{ExtractGt}
\title{A function to extract a genotype from a vcfR::VCF object.}
\usage{
ExtractGt(vcf, min.depth = NULL, min.sample.qual = NULL,
  min.variant.qual = NULL, verbose = TRUE)
}
\arguments{
\item{vcf}{A vcfR::VCF object}

\item{min.depth}{Optional. A threshold depth under which the data will be
censored (possibly a positive Integer)}

\item{min.sample.qual}{Optional. A threshold ratio under which the samples
with too much missing data will be removed (from 0 to 1)}

\item{min.variant.qual}{Optional. A threshold ratio under which the variants
with too much missing data will be removed (from 0 to 1)}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution}
}
\value{
A list of dataframes. \code{list$gt} is the genotype dataframe, \code{list$dp} is the
depth dataframe.
}
\description{
A function to extract a genotype from a vcfR::VCF object.
}
\examples{
ExtractGt(vcf)
ExtractGt(vcf, min.sample.qual = 0.65, min.variant.qual = 0.4)
ExtractGt(vcf, min.depth = 4, verbose = FALSE)
}
