% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{FindIdsAlFreqs}
\alias{FindIdsAlFreqs}
\title{A function to find the ids of freq.al columns in a frequency matrix.}
\usage{
FindIdsAlFreqs(freq)
}
\arguments{
\item{freq}{the realtive frequency matrix}
}
\value{
A list of integers containing ids of specific columns. Items names are :
"ref", "alt", "missval". The value are the corresponding column ids.
}
\description{
A function to find the ids of freq.al columns in a frequency matrix.
}
\examples{
FindIdsAlFeqs(freq)
}
