% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{FindIdsGtCounts}
\alias{FindIdsGtCounts}
\title{A function to find the ids of counts.gt columns in a frequency matrix.}
\usage{
FindIdsGtCounts(freq)
}
\arguments{
\item{freq}{the absolute frequency matrix}
}
\value{
A list of integers containing ids of specific columns. Items names are :
"homo.ref", "hetero", "homo.alt", "missval", "total". The value are the
corresponding column ids.
}
\description{
A function to find the ids of counts.gt columns in a frequency matrix.
}
\examples{
\dontrun{
FindIdsGtCounts(freq)
}
}
