% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{FindSH}
\alias{FindSH}
\title{A function to retrieve s and h fitness parameters from benchmark results.}
\usage{
FindSH(benchmark)
}
\arguments{
\item{benchmark}{Results of \code{link{Benchmark}}}
}
\value{
Return the mod of estimated s and h fitness parameters from benchmark
results.
}
\description{
A function to retrieve s and h fitness parameters from benchmark results.
}
\seealso{
\code{\link{Benchmark}} needed for \code{benchmark} parameter.
}
