% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{IdentifyGt}
\alias{IdentifyGt}
\title{A function to classify samples by genotypes for a single variant, use regex.
NOT USED ANYMORE.}
\usage{
IdentifyGt(variant)
}
\arguments{
\item{variant}{a variant character data frame.}
}
\value{
A list of vectors which contain the ids of samples for each genotype.
}
\description{
A function to classify samples by genotypes for a single variant, use regex.
NOT USED ANYMORE.
}
\examples{
\dontrun{
 IdentifyGt(variant)
}
}
