% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{IsAtLeastNARatio}
\alias{IsAtLeastNARatio}
\title{A function to check if the ratio of NA in a vector is over a threshold.}
\usage{
IsAtLeastNARatio(vect, min.ratio)
}
\arguments{
\item{vect}{A vector to analyse}

\item{min.ratio}{A threshold ratio of NA values in vect (from 0 to 1)}
}
\value{
Logical. TRUE if the ratio of NA in vect is at least minRatio, else FALSE.
}
\description{
A function to check if the ratio of NA in a vector is over a threshold.
}
\examples{
IsAboveNARatio(vector, 0.2)
}
