% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{IterRandPick}
\alias{IterRandPick}
\title{A function to iterate significant snps counting in a genotype matrx with a
random permutation of the survival vector}
\usage{
IterRandPick(gt, survival, max.p.neutral = 0.1, iter = 10,
  odded.pos = NULL, verbose = TRUE)
}
\arguments{
\item{gt}{A genotype matrix}

\item{survival}{A survival vector to shuffle}

\item{max.p.neutral}{A threshold for \code{p.neutral} under which the snp
will be counted (default is 0.01)}

\item{iter}{Number of iteration (default is 10)}

\item{odded.pos}{Optional. Reference to a variant in \code{gt} as a string.
Iterate random pick according to the observed odds of genotype at this
position in the genome.}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
An integer vector. Each item is the number of snps with low \code{p.neutral}
value in \code{gt} data frame after a random permutation of the survival
vector. Countains \code{iter} items.
}
\description{
A function to iterate significant snps counting in a genotype matrx with a
random permutation of the survival vector
}
\examples{
IterRandomPick(gt, survival)
IterRandomPick(gt, survival, max.p.neutral = 0.05)
IterRandomPick(gt, survival, iter = 1000, verbose = FALSE)
}
\seealso{
For more detail, see \code{link{PermutSurv}} and
\code{link{CountSignSnps}} which this function bind.
}
