% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{LoadVcf}
\alias{LoadVcf}
\title{A function to get a filtered (and possibly splitted) vcfR::VCF object from
a VCF file.}
\usage{
LoadVcf(file.path, survivals = NULL, only.biallelic = TRUE,
  only.snp = TRUE, verbose = TRUE)
}
\arguments{
\item{file.path}{The path to the VCF file you want to load}

\item{survivals}{Optional. A logical vector for splitting the data into two
vcfR::VCF objects. The length of the vector must match the number of samples
in the VCF file.}

\item{only.biallelic}{Logical. If TRUE (default), only biallelic variants
will be loaded.}

\item{only.snp}{Logical. If TRUE (default), the indel variants will be
omited.}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A vcR:VCF object or a list of vcfR:VCF objects if a survivals vector was
provided. 
If a list is returned, list$alive is the vcfR::VCF object containing the
samples which the position corresponded with a TRUE in the survivals vector.
The other samples are contained in list$dead as an other vcfR::VCF object.
}
\description{
A function to get a filtered (and possibly splitted) vcfR::VCF object from
a VCF file.
}
\examples{
LoadVCF("example.vcf")
LoadVCF("example.vcf", survivals = c(TRUE, FALSE, FALSE, TRUE, ...))
LoadVCF("example.vcf", survivals = mySurvVector, verbose = FALSE)
LoadVCF("example.vcf", only.biallelic = FALSE, only.snp = FALSE)
}
\seealso{
For more information, see \code{\link{SplitVcf}},
\code{\link{RmNonBiallelics}} and \code{\link{RmIndels}} which this
function bind.
}
