% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{PermutSurv}
\alias{PermutSurv}
\title{A function to calculate probabilities of neutral model and positive
selection model from a genotype matrix whit a randomly shuffled survival
vector}
\usage{
PermutSurv(gt, survival, odds = NULL, verbose = TRUE)
}
\arguments{
\item{gt}{A genotype matrix}

\item{survival}{A survival vector to shuffle}

\item{odds}{Optional. Not used yet.}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A data frame of frequencies and probabilities, one snp per row.
}
\description{
A function to calculate probabilities of neutral model and positive
selection model from a genotype matrix whit a randomly shuffled survival
vector
}
\examples{
PermutSurv(gt, c(TRUE, TRUE, FALSE, ...))
PermutSurv(gt, survival, verbose = FALSE)
}
\seealso{
For more detail, see \code{link{AnalyseExpt}} which this function
bind.
}
