% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{RmIndels}
\alias{RmIndels}
\title{A function to remove indel variants from a vcfR::VCF object}
\usage{
RmIndels(vcf, verbose = TRUE)
}
\arguments{
\item{vcf}{The original vcfR::VCF object to process}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A vcfR::VCF object rid of its indel variants
}
\description{
A function to remove indel variants from a vcfR::VCF object
}
\examples{
RmIndels(vcf)
RmIndels(vcf, verbose = FALSE)
}
