% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{RmNonBiallelics}
\alias{RmNonBiallelics}
\title{A function to remove non-biallelic variants from a vcfR::VCF object}
\usage{
RmNonBiallelics(vcf, verbose = TRUE)
}
\arguments{
\item{vcf}{The original vcfR::VCF object to process}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A vcfR::VCF object rid of its non-biallelic variants
}
\description{
A function to remove non-biallelic variants from a vcfR::VCF object
}
\examples{
\dontrun{
RmNonBiallelics(vcf)
RmNonBiallelics(vcf, verbose = FALSE)
}
}
