% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{ShapeAsVariantPos}
\alias{ShapeAsVariantPos}
\title{A function to find the ranking of a variant in a list of ordered variants.}
\usage{
ShapeAsVariantPos(ordered.variants, variant.name)
}
\arguments{
\item{ordered.variants}{A list of ordered variants name.}

\item{variant.name}{The name of a variant to look for.}
}
\value{
The rank of \code{variant.name} in \code{ordered.variant}.
}
\description{
A function to find the ranking of a variant in a list of ordered variants.
}
\seealso{
\code{\link{Benchmark}} which binds this function.
}
