% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SliceDfRows}
\alias{SliceDfRows}
\title{A function to slice data frames without loosing row names}
\usage{
SliceDfRows(df, rows)
}
\arguments{
\item{df}{A dataframe to slice}

\item{rows}{A integer vector. The ids of the rows to return.}
}
\value{
A dataframe containing only the rows number \code{rows} from \code{df}.
}
\description{
A function to slice data frames without loosing row names
}
\examples{
SliceDfRows(df, 1)
SliceDfRows(df, 1:5)
SliceDfRows(df, c(1,3,5))
}
