% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{SortSamplesVariant}
\alias{SortSamplesVariant}
\title{A function to sort occurences of genotypes in multiple combined variant
vectors.}
\usage{
SortSamplesVariant(variants, group.nas = FALSE)
}
\arguments{
\item{variants}{A vector of variants, subset of a genotype matrix.}

\item{group.nas}{If FALSE (default), each combinations with missing values
will be considered on its own. Else, combinations with missing values will
be grouped as one category.}
}
\value{
A list of vectors which contains the ids of categorized samples.
}
\description{
A function to sort occurences of genotypes in multiple combined variant
vectors.
}
\examples{
\dontrun{
SortSamplesMultiVariant(variants, group.nas = TRUE)
}
}
\seealso{
\code{\link{CalcFreqMultiVariant}} which binds this function
}
