% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{SplitGt}
\alias{SplitGt}
\title{A function to split a genotype matrix two genotype matrix according to a
logical vector}
\usage{
SplitGt(gt, survival, verbose = TRUE)
}
\arguments{
\item{gt}{The original genotype matrix to split}

\item{survival}{A logical vector for splitting the data into two gt matrix.
The length of the vector must match the number of samples in the gt matrix.}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A list of genotype matrix list.
\code{list$alive} is the gt matrix containing the samples which position
corresponds with a TRUE in the \code{survival} vector.
The other samples are contained in \code{list$dead} as an other gt matrix
}
\description{
A function to split a genotype matrix two genotype matrix according to a
logical vector
}
\examples{
SplitGt(gt, c(TRUE, TRUE, FALSE, ...))
SplitGt(gt, survival.vector, verbose = FALSE)
}
