% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{SplitVcf}
\alias{SplitVcf}
\title{A function to split a vcfR::VCF object in two vcfR::VCF object according to
a logical vector}
\usage{
SplitVcf(vcf, survival, verbose = TRUE)
}
\arguments{
\item{vcf}{The original vcfR::VCF object to split}

\item{survival}{A logical vector for splitting the data into two vcfR::VCF
objects.
The length of the vector must match the number of samples in the VCF file.}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A list of vcfR:VCF objects.
\code{list$alive} is the vcfR::VCF object containing the samples which position
corresponds with a TRUE in the \code{survival} vector.
The other samples are contained in \code{list$dead} as an other vcfR::VCF object.
}
\description{
A function to split a vcfR::VCF object in two vcfR::VCF object according to
a logical vector
}
\examples{
SplitVcf(vcf, c(TRUE, TRUE, FALSE, ...))
SplitVcf(vcf, survival.vector, verbose = FALSE)
}
