% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SplitVect}
\alias{SplitVect}
\title{A function to split a vector into two vectors according to a survival 
logical vector}
\usage{
SplitVect(vect, survival, verbose = TRUE)
}
\arguments{
\item{vect}{The original vector to split}

\item{survival}{A logical vector for splitting the data into two vectors.
The length of \code{survival} must match the length of \code{vector}}

\item{verbose}{Logical. If TRUE (default), report status of the process
along the execution.}
}
\value{
A list of vectors.
\code{list$alive} is the vector containing the samples which position
corresponds with a TRUE in the \code{survival} vector.
The other samples are contained in \code{list$dead} as an other vector
}
\description{
A function to split a vector into two vectors according to a survival 
logical vector
}
\examples{
\dontrun{
SplitGt(gt, c(TRUE, TRUE, FALSE, ...))
SplitGt(gt, survival.vector, verbose = FALSE)
}
}
