% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{txtTimerBar}
\alias{txtTimerBar}
\title{Text progress bar with time.}
\usage{
txtTimerBar(n = 1)
}
\description{
A textual progress bar that estimates time remaining. It displays the
estimated time remaining and, when finished, total duration.  Lifted from the
plyr package.  Update with \code{setTxtProgressBar}
}
