% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{AnalyseExpt}
\alias{AnalyseExpt}
\title{A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.}
\usage{
AnalyseExpt(gt, survival, population = NULL, min.freq.al = NULL,
  location.cols = TRUE, deltas = TRUE, p.values = TRUE,
  genotypic = FALSE, backup.path = NULL)
}
\arguments{
\item{gt}{A genotype data frame of your original population}

\item{survival}{A logical vector. TRUE survived the experiment, FALSE died}

\item{population}{A logical vector. TRUE belong to pop1, FALSE belong to
pop2. A M&C FST estimation will be included in the result.}

\item{min.freq.al}{Optional. Numeric from 0 to 1. Variants with MAF under
this threshold will be ommited.}

\item{location.cols}{If TRUE, adds two columns to the result, which are the
scaffold name and the locus name as factors (deafult is FALSE)}

\item{deltas}{If TRUE (default), include columns of differencies between
relative allelic frequencies after selection and before selection in the
result. Positive value means that the allele is more common in the survivers
population.}

\item{p.values}{If TRUE (default), some probabilities will be included in
the result, as well as frequencies}

\item{genotypic}{If TRUE, will include counts of genotype in the result.}

\item{backup.path}{Optionnal. A path where backup files can be stored}
}
\value{
A data frame of frequencies and probabilities from your experiment.
}
\description{
A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.
}
\examples{
AnalyseExpt(gt, survival)
AnalyseExpt(gt, survival, min.freq.al = 0.1, location.cols = FALSE)
AnalyseExpt(gt, survival, deltas = FALSE, p.values = FALSE)
AnalyseExpt(gt, survival, backup.path = "example.csv")
}
\seealso{
For more information, see \code{\link{AnalyseSplittedExpt}} which
this function bind.
}
