% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{AnalyseExpt}
\alias{AnalyseExpt}
\title{A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.}
\usage{
AnalyseExpt(gt.alive, gt.dead, location.cols = TRUE, p.values = TRUE,
  backup.path = NULL)
}
\arguments{
\item{gt.alive}{A genotype data frame of the alive population}

\item{gt.dead}{A genotype data frame of the dead population}

\item{location.cols}{If TRUE, adds two columns to the result, which are the
scaffold name and the locus name as factors (deafult is FALSE)}

\item{p.values}{If TRUE (default), some probabilities will be included in
the result, as well as frequencies}

\item{backup.path}{Optionnal. A path where backup files can be stored}
}
\value{
A data frame of frequencies and probabilities from your experiment.
}
\description{
A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.
}
\examples{
AnalyseExpt(gt.alive, gt.dead)
AnalyseExpt(gt.alive, gt.dead, p.values = FALSE)
AnalyseExpt(gt.alive, gt.dead, backup.path = "example.csv")
}
\seealso{
For more information, see \code{\link{CalcProbsSelection}} and
\code{\link{CalcFreqGt}} which this function bind.
}
