% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{AnalyseExpt}
\alias{AnalyseExpt}
\title{A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.}
\usage{
AnalyseExpt(gt.alive, gt.dead, p.values = TRUE, backup.path = NULL)
}
\arguments{
\item{gt.alive}{A genotype data frame of the alive population}

\item{gt.dead}{A genotype data frame of the dead population}

\item{p.values}{If TRUE (default), some probabilities will be included in
the result, as well as frequencies}

\item{backup.path}{Optionnal. A path where backup files can be stored}
}
\value{
A data frame of frequencies and probabilities from your experiment.
}
\description{
A function to analyse a selection experiment, including observed frequencies
of genotype and probabilities that they are due to random picking and
positive selection.
}
\examples{
Analyse 
}
\seealso{
For more information, see \code{\link{CalcProbsSelection}} and
\code{\link{FreqGt}} which this function bind.
}
