% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{TxtTimerBar}
\alias{TxtTimerBar}
\title{A function to create a text progress bar.}
\usage{
TxtTimerBar(n = 1)
}
\arguments{
\item{n}{Maximum value of progression.}
}
\value{
Print a progress bar, use .update(progress) to update and .kill() to kill.
}
\description{
A function to create a text progress bar.
}
